+++ Utils/Math/MyMath.cpp
@@ -2,16 +2,16 @@
 
 #include<math.h>
 
-Vector3d Vector3d::Cross(const Vector3d& v) const
+Vector3f Vector3f::Cross(const Vector3f& v) const
 {
 	const double _x = y * v.z - z * v.y;
 	const double _y = z * v.x - x * v.z;
 	const double _z = x * v.y - y * v.x;
 
-	return Vector3d(_x, _y, _z);
+	return Vector3f(_x, _y, _z);
 }
 
-Vector3d& Vector3d::Normalize()
+Vector3f& Vector3f::Normalize()
 {
 	const double Length = sqrt(x * x + y * y + z * z);
